hyprland: remove
https://web.archive.org/web/20231106052720/https://blog.vaxry.net/articles/2023-inclusiveActivists

I don't like this.

Personally, I don't agree with the justification given here. Yes, we're all people, but respectfully, I don't agree with people needing to be strong and tough to be on the internet.

Also, in my opinion, there were way better ways of dealing with this than condemnation and projection of frustrations.

The internet lets people be who they are while having control over what they share of their real lives. It leads to great things, like the socially anxious experimenting with their identity, and terrible things, like cyberbullying and hate speech.

It's hard to be nice to people. But some people don't even try, and that's sad.

In my opinion, it's harder to listen and be open to change rather than tough and resilient. Being tough all the time is not healthy.

I'm glad to see that there's a bit of an effort in the CoC, though there's seems to be a vocal part of the group who still pushes for no rules. (issue 3366 on hyprland)

I firmly believe that, an inclusive community cannot be tolerant of all behaviour. That's what the rules are made for - punishing by the basis of non-inclusive behaviour, and not by the people themselves. I feel a lot of people conflate that.

We don't exclude people because they are people, we never should. But when people hurt others, that should be taken seriously. Sometimes that even means giving consequences to the people who do these practices.

Vaxry, if you do see this somehow - let's have a chat. I want to hear your story. But, looking at the community right now, I don't feel comfortable being a part of it. Sorry.